feat: adding FlightController and refactoring all main code

+++ src/main/java/com/flightontime/api/dto/PredictionResponseDTO.java
@@ -5,6 +5,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public record PredictionResponseDTO(
 
     @JsonProperty("previsao_atraso") Integer prediction,
-    @JsonProperty("probabilidade_atraso") Double probability
+    @JsonProperty("probabilidade_atraso") Float probability
 ) {
 }
